requirements.txt
python-dotenv
huggingface_hub
transformers
datasets